Conflicts:
	aZynEManager/aZynEManager/frmDistributor.cs

+++ aZynEManager/Paradiso/SeatingPage.xaml.cs
@@ -149,7 +149,8 @@ namespace Paradiso
             if (window != null)
                 window.KeyDown -= Page_PreviewKeyDown;
 
-            NavigationService.Navigate(new Uri("MovieCalendarPage.xaml", UriKind.Relative));
+            if (NavigationService != null) 
+                NavigationService.Navigate(new Uri("MovieCalendarPage.xaml", UriKind.Relative));
         }
 
 
@@ -193,7 +194,8 @@ namespace Paradiso
                     if (window != null)
                         window.KeyDown -= Page_PreviewKeyDown;
 
-                    NavigationService.Navigate(new Uri("MovieCalendarPage.xaml", UriKind.Relative));
+                    if (NavigationService != null) 
+                        NavigationService.Navigate(new Uri("MovieCalendarPage.xaml", UriKind.Relative));
                     return;
                 }
 
@@ -212,7 +214,8 @@ namespace Paradiso
                     if (window != null)
                         window.KeyDown -= Page_PreviewKeyDown;
 
-                    NavigationService.Navigate(new Uri("MovieCalendarPage.xaml", UriKind.Relative));
+                    if (NavigationService != null) 
+                        NavigationService.Navigate(new Uri("MovieCalendarPage.xaml", UriKind.Relative));
                     return;
                 }
 
@@ -628,7 +631,8 @@ namespace Paradiso
                 if (window != null)
                     window.KeyDown -= Page_PreviewKeyDown;
 
-                NavigationService.Navigate(new Uri("TenderAmountPage.xaml", UriKind.Relative));
+                if (NavigationService != null) 
+                    NavigationService.Navigate(new Uri("TenderAmountPage.xaml", UriKind.Relative));
             }
         }
 
@@ -975,6 +979,7 @@ namespace Paradiso
                     {
                         try
                         {
+                            string strSessionId = ParadisoObjectManager.GetInstance().SessionId;
                             foreach (ReservedSeatModel rs in rsl.ReservedSeats)
                             {
                                 var _rs = (from mslhs in context.movies_schedule_list_house_seat
@@ -982,7 +987,16 @@ namespace Paradiso
                                            select mslhs).FirstOrDefault();
                                 if (_rs != null)
                                 {
-                                    context.movies_schedule_list_house_seat.DeleteObject(_rs);
+                                    if (ParadisoObjectManager.GetInstance().IsReserveUnreservedSeats)
+                                    {
+                                        _rs.session_id = strSessionId;
+                                        _rs.reserved_date = ParadisoObjectManager.GetInstance().CurrentDate;
+                                    }
+                                    else
+                                    {
+                                        context.movies_schedule_list_house_seat.DeleteObject(_rs);
+                                    }
+                                    
                                     context.SaveChanges();
                                 }
                             }
